chore: move the Multiaddr wrappers to their own module
